More metadata and faster plots Bedmap + ETOPO1
Include units and long names for coordinates and data variables for
Bedmap2 and ETOPO1. xarray puts these in the plots automatically, which
is nice.
Speed up the plots by not using projections. Not ideal but they look
better and run faster. Includes a notice not to do that in practice.