Use pooch.os_cache to determine the data location
The function finds the cache folder for each operating system. This is
more user friendly than setting the directory by hand (it plays nicely
with backups and cleaning programs). Can still be overwritten by the
environment variable. Provide a function that returns the cache location
for convenience to the user as well.