Add Slab2 model to Rockhound (#62)
Add function to fetch and load data from Slab2: a geometry model of
subduction zones, provided by the USGS. The model contains gridded data
for 27 subduction zones describing the depth, thickness, dip and strike
angles of the subducted slab. The data for each subduction zone is
fetched and loaded into a xarray.Dataset. The depth and thickness are
converted to meters and the coordinates renamed to "longitude" and
"latitude". Add metadata to the Dataset like the datum and the doi of
the model. Add netcdf4 as a requirement for Rockhound.