Add Release Drafter to generate the changelog (#88)
This Github Action automatically generates a changelog based on pull
requests merged. The `.github/workflows/release-drafter.yml` sets up the
workflow and `.github/release-drafter.yml` configures the generated
changelog. A caveat is that the names of contributors is given only as
their Github handle and not the full names (which we were using before).